fix:field ipk table mahasiswa change to decimal

+++ database/migrations/2025_03_24_081317_create_mahasiswas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('nama');
             $table->string('nim')->unique();
             $table->integer('semester');
-            $table->integer('ipk');
+            $table->decimal('ipk', 3, 2);
             $table->timestamps();
         });
     }
